HUD: Updated GameLevel to spawn the HUD widget. Created HUD widget that contains references of HUD_Ammo, HUD_Health, and HUD_Progress. Created HUD_Ammo widget with function that updates the ammo amount visuals. Created HUD_Health widget with function that updates the health bar visuals. Created HUD_Progress widget with function that sets the progress bar visuals and function that updates the progress of the bar. Created ProgressBarDataTable, ProgressBarStructure, and ProgressBarStyleEnum to easily update the different ProgressBar styles. Added YouMurdererBB font.
BPs:
Created Candle BP with function that lights the candle on/off.
Created Candle_Manager BP with 5 instances of the Candle BP and a fanction that creates an array of the 5 candles and lights the next available candle every time it gets called.
Created Sould BP with tik function that interpolates the position of the actor to the next candle.
Created CandleFire and SoulParticle niagara particles.